+++ frontend/wailsjs/go/models.ts
@@ -439,7 +439,6 @@ export namespace model {
 	    category: string;
 	    exists: boolean;
 	    comment: string;
-	    fullName: string;
 	    parents: string[];
 	    attributes: Record<string, boolean>;
 	    settings: Record<string, string>;
@@ -459,7 +458,6 @@ export namespace model {
 	        this.category = source["category"];
 	        this.exists = source["exists"];
 	        this.comment = source["comment"];
-	        this.fullName = source["fullName"];
 	        this.parents = source["parents"];
 	        this.attributes = source["attributes"];
 	        this.settings = source["settings"];
@@ -475,7 +473,6 @@ export namespace model {
 	    category: string;
 	    loginName: string;
 	    comment: string;
-	    fullName: string;
 	
 	    static createFrom(source: any = {}) {
 	        return new RoleMatch(source);
@@ -489,7 +486,6 @@ export namespace model {
 	        this.category = source["category"];
 	        this.loginName = source["loginName"];
 	        this.comment = source["comment"];
-	        this.fullName = source["fullName"];
 	    }
 	}
 	export class ClusterRoleMatches {
@@ -594,6 +590,20 @@ export namespace model {
 	        if ('string' === typeof source) source = JSON.parse(source);
 	        this.categoryIds = source["categoryIds"];
 	        this.clusterIds = source["clusterIds"];
+	    }
+	}
+	export class SearchColumn {
+	    label: string;
+	    template: string;
+	
+	    static createFrom(source: any = {}) {
+	        return new SearchColumn(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.label = source["label"];
+	        this.template = source["template"];
 	    }
 	}
 	export class PasswordGen {
@@ -778,6 +788,7 @@ export namespace model {
 	    ui: UISettings;
 	    parentRoles: string[];
 	    commentFields: CommentField[];
+	    searchColumns: SearchColumn[];
 	    targets: TargetSelection;
 	    windowWidth?: number;
 	    windowHeight?: number;
@@ -798,6 +809,7 @@ export namespace model {
 	        this.ui = this.convertValues(source["ui"], UISettings);
 	        this.parentRoles = source["parentRoles"];
 	        this.commentFields = this.convertValues(source["commentFields"], CommentField);
+	        this.searchColumns = this.convertValues(source["searchColumns"], SearchColumn);
 	        this.targets = this.convertValues(source["targets"], TargetSelection);
 	        this.windowWidth = source["windowWidth"];
 	        this.windowHeight = source["windowHeight"];
@@ -1033,6 +1045,7 @@ export namespace model {
 		    return a;
 		}
 	}
+